Initial V1.0 release
This was designed in 2016 but never officially released. I have
corrected a few things and saved the files in the latest Kicad v6.
Gerbers are included. The schematic design was verified working, however
the inital pcb had an error with the power connector orientation. That
has been fixed. Gerbers have not been verified so USE AT OWN RISK!